feat: add email and file upload tasks for Celery
